Bug 1043112 - Add support for opening core files to run-gdb wrapper. r=dhylands

+++ run-gdb.sh
@@ -48,38 +48,40 @@ GDBINIT=/tmp/b2g.gdbinit.$(whoami).$$
 GONK_OBJDIR=out/target/product/$DEVICE
 SYMDIR=$GONK_OBJDIR/symbols
 
-GDBSERVER_PID=$(get_pid_by_name gdbserver)
+if [ "$1" != "core" ] ; then
+   GDBSERVER_PID=$(get_pid_by_name gdbserver)
 
-GDB_PORT=$((10000 + $(id -u) % 50000))
-if [ "$1" = "vgdb"  -a  -n "$2" ] ; then
-   GDB_PORT="$2"
-elif [ "$1" = "attach"  -a  -n "$2" ] ; then
-   B2G_PID=$2
-   if [ -z "$($ADB ls /proc/$B2G_PID)" ] ; then
-      ATTACH_TARGET=$B2G_PID
-      B2G_PID=$(get_pid_by_name "$B2G_PID")
-      if [ -z "$B2G_PID" ] ; then
-        echo Error: PID $ATTACH_TARGET is invalid
-        exit 1;
+   GDB_PORT=$((10000 + $(id -u) % 50000))
+   if [ "$1" = "vgdb"  -a  -n "$2" ] ; then
+      GDB_PORT="$2"
+   elif [ "$1" = "attach"  -a  -n "$2" ] ; then
+      B2G_PID=$2
+      if [ -z "$($ADB ls /proc/$B2G_PID)" ] ; then
+         ATTACH_TARGET=$B2G_PID
+         B2G_PID=$(get_pid_by_name "$B2G_PID")
+         if [ -z "$B2G_PID" ] ; then
+           echo Error: PID $ATTACH_TARGET is invalid
+           exit 1;
+         fi
+         echo "Found $ATTACH_TARGET PID: $B2G_PID"
       fi
-      echo "Found $ATTACH_TARGET PID: $B2G_PID"
+      GDB_PORT=$((10000 + ($B2G_PID + $(id -u)) % 50000))
+      # cmdline is null separated
+      B2G_BIN=$($ADB shell cat /proc/$B2G_PID/cmdline | tr '\0' '\n' | head -1)
+   else
+      B2G_PID=$(get_pid_by_name b2g)
    fi
-   GDB_PORT=$((10000 + ($B2G_PID + $(id -u)) % 50000))
-   # cmdline is null separated
-   B2G_BIN=$($ADB shell cat /proc/$B2G_PID/cmdline | tr '\0' '\n' | head -1)
-else
-   B2G_PID=$(get_pid_by_name b2g)
+
+   for p in $GDBSERVER_PID ; do
+      $ADB shell cat /proc/$p/cmdline | grep -q :$GDB_PORT && ( \
+         echo ..killing gdbserver pid $p
+         $ADB shell kill $p
+      ) || echo ..ignoring gdbserver pid $p
+
+   done
+
+   $ADB forward tcp:$GDB_PORT tcp:$GDB_PORT
 fi
-
-for p in $GDBSERVER_PID ; do
-   $ADB shell cat /proc/$p/cmdline | grep -q :$GDB_PORT && ( \
-      echo ..killing gdbserver pid $p
-      $ADB shell kill $p
-   ) || echo ..ignoring gdbserver pid $p
-
-done
-
-$ADB forward tcp:$GDB_PORT tcp:$GDB_PORT
 
 if [ "$1" = "attach" ]; then
    if [ -z "$B2G_PID" ]; then
@@ -88,6 +90,23 @@ if [ "$1" = "attach" ]; then
    fi
 
    $ADB shell gdbserver :$GDB_PORT --attach $B2G_PID &
+elif [ "$1" == "core" ]; then
+   if [ -z "$3" ]; then
+     CORE_FILE=$2
+   else
+     B2G_BIN=$2
+     CORE_FILE=$3
+   fi
+
+   if [ "$B2G_BIN" == "" -o "$CORE_FILE" == "" ]; then
+     echo "Usage: $SCRIPT_NAME core [bin] <core>"
+     exit 1
+   fi
+
+   if [ ! -f $CORE_FILE ]; then
+     echo "Error: $CORE_FILE not found."
+     exit 1
+   fi
 elif [ "$1" != "vgdb" ]; then
    if [ -n "$1" ]; then
       B2G_BIN=$1
@@ -118,12 +137,20 @@ echo "handle SIGPIPE nostop" >> $GDBINIT
 echo "set solib-search-path $GECKO_OBJDIR/dist/bin:$SYMDIR/system/lib:$SYMDIR/system/lib/hw:$SYMDIR/system/lib/egl:$SYMDIR/system/bin:$GONK_OBJDIR/system/lib:$GONK_OBJDIR/system/lib/egl:$GONK_OBJDIR/system/lib/hw:$GONK_OBJDIR/system/vendor/lib:$GONK_OBJDIR/system/vendor/lib/hw:$GONK_OBJDIR/system/vendor/lib/egl" >> $GDBINIT
 if [ "$1" == "vgdb" ] ; then
   echo "target remote :$GDB_PORT" >> $GDBINIT
-else
+elif [ "$1" != "core" ]; then
   echo "target extended-remote :$GDB_PORT" >> $GDBINIT
 fi
 
 PROG=$GECKO_OBJDIR/dist/bin/$(basename $B2G_BIN)
-[ -f $PROG ] || PROG=${SYMDIR}${B2G_BIN}
+[ -f $PROG ] || PROG=${SYMDIR}/${B2G_BIN}
+[ -f $PROG ] || PROG=${B2G_BIN}
+if [ ! -f $PROG ]; then
+  echo "Error: program to debug not found:"
+  echo "  $GECKO_OBJDIR/dist/bin/$(basename $B2G_BIN)"
+  echo "  $SYMDIR/$B2G_BIN"
+  echo "  $B2G_BIN"
+  exit 1
+fi
 
 if [[ "$-" == *x* ]]; then
     # Since we got here, set -x was enabled near the top of the file. print
@@ -134,9 +161,9 @@ if [[ "$-" == *x* ]]; then
 fi
 
 if [ "$SCRIPT_NAME" == "run-ddd.sh" ]; then
-    echo "ddd --debugger \"$GDB -x $GDBINIT\" $PROG"
-    ddd --debugger "$GDB -x $GDBINIT" $PROG
+    echo "ddd --debugger \"$GDB -x $GDBINIT\" $PROG $CORE_FILE"
+    ddd --debugger "$GDB -x $GDBINIT" $PROG $CORE_FILE
 else
-    echo $GDB -x $GDBINIT $PROG
-    $GDB -x $GDBINIT $PROG
+    echo $GDB -x $GDBINIT $PROG $CORE_FILE 
+    $GDB -x $GDBINIT $PROG $CORE_FILE
 fi
